requirements.txt
mysql-connector-python==8.2.0
tabulate==0.9.0
python-dotenv==1.0.0
requests==2.31.0
urllib3==2.2.0